backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.2
flask-cors==6.0.2
flask-limiter==3.8.0
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.4
zipp==3.23.0
pymongo
python-dotenv
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
razorpay==1.4.2
gunicorn==21.2.0
redis==5.0.8
limits>=3.0
